+++ .changeset/small-papayas-laugh.md
@@ -1,0 +1,5 @@
+---
+'svelte': patch
+---
+
+breaking: remove Component type, keep using SvelteComponent instead
