+++ packages/svelte/src/main/ambient.d.ts
@@ -1,5 +1,5 @@
 declare module '*.svelte' {
-	export { Component as default } from 'svelte';
+	export { SvelteComponent as default } from 'svelte';
 }
 
 /**
